fix mac build
